requirements.txt
selenium
pydantic
email_listener
pyyaml
pypdf2